Move finding to toolbar controller, finding works across all tabs now
