+++ proj9DouglasHanssenMacDonaldZhang/Controllers/EditMenuController.java
@@ -30,14 +30,12 @@ import java.util.ArrayList;
 public class EditMenuController
 {
     private CodeAreaTabPane tabPane;
-    private TextField textField;
 
     /**
      * Constructor for the Edit Menu Controller
      */
-    public EditMenuController(CodeAreaTabPane tabPane, TextField textField) {
+    public EditMenuController(CodeAreaTabPane tabPane) {
         this.tabPane = tabPane;
-        this.textField = textField;
     }
 
     /**
@@ -252,73 +250,6 @@ public class EditMenuController
         alert.setHeaderText("Checking brackets, parentheses and braces");
         alert.setContentText(bracesMessage + parensMessage + bracketsMessage);
         alert.showAndWait();
-    }
-
-    /**
-     * Finds text in the current code area that matches the text in the
-     * textField. Is case-sensitive.
-     */
-    public void handleFind() {
-        CodeArea curCodeArea = this.tabPane.getCurCodeArea();
-        this.textField.setVisible(true);
-        this.textField.requestFocus();
-        this.textField.setOnKeyPressed(keyEvent -> {
-            KeyCode keyCode = keyEvent.getCode();
-            String searchedText = this.textField.getText();
-
-            // check for enter key and find search results
-            if(keyCode.equals(KeyCode.ENTER)) {
-                resetHighlighting(curCodeArea);
-
-                ArrayList<Integer> matchIndices = this.matchesString(searchedText);
-                if(matchIndices != null) {
-                    for (Integer index : matchIndices) {
-                        curCodeArea.setStyleClass(index, index + searchedText.length(), "find");
-                    }
-                }
-            }
-            // check if the search field is empty, reset text
-            // this currently is delayed by one character
-            if(searchedText.trim().isEmpty()) {
-                resetHighlighting(curCodeArea);
-            }
-        });
-    }
-
-    /**
-     * Helper method for handleFind
-     * Finds matching text in the CodeArea by getting the index
-     * of the text that matches the search input.
-     * @param input
-     * @return ArrayList of type Integer with indices of matching text
-     */
-    private ArrayList<Integer> matchesString(String input) {
-        ArrayList<Integer> inputArray = new ArrayList<Integer>();
-        CodeArea curCodeArea = this.tabPane.getCurCodeArea();
-
-        if(input.length() == 0) {
-            return null;
-        }
-
-        if (curCodeArea.getText().contains(input)) {
-            int index = curCodeArea.getText().indexOf(input);
-            while (index != -1) {
-                inputArray.add(index);
-                index = curCodeArea.getText().indexOf(input, index + 1);
-            }
-        }
-        return inputArray;
-    }
-
-    /**
-     * Reset highlighting for current code area
-     * @param curCodeArea is the current code area being viewed in the tab pane
-     */
-    private void resetHighlighting(CodeArea curCodeArea) {
-        // reset the highlighting
-        curCodeArea.setStyleClass(0, curCodeArea.getText().length(), "reset-found-words");
-        // re-compute the text styling, because it gets removed when found words is reset
-        curCodeArea.setStyleSpans(0, JavaCodeArea.computeHighlighting(curCodeArea.getText()));
     }
 
     /**
